Try to preserve mtime on deploys.
We'll first try to "git checkout origin/<branch>", which will preserve
the mtime of files that haven't changed. If that fails, we'll fall
back to "git reset --hard origin/<branch>", which is not
mtime-preserving.

If the user doesn't screw around with the git repo in cached-copy,
then we shouldn't need to fall back on a hard reset. It's mainly there
in case the user does something to that repository that causes a
simple checkout to fail (locally modified files, e.g.).

Further, it's now possible to deploy tags and SHAs using eysd.

Also added more verbose output to better show what's being run on the
system. This made it much easier to see what went wrong.

[#954 state:merged]
[#1456]

Change-Id: I90d455d21b02f803589a911294d8e29b177b2dbf
Reviewed-on: http://review.engineyard.com/244
Reviewed-by: Andy Delcambre <adelcambre@engineyard.com>
Tested-by: Andy Delcambre <adelcambre@engineyard.com>