Make EY::Server.all start off with localhost
This way, you don't have to specify 'localhost' or some dummy string
in the --instance argument just to specify roles for localhost.

This does not address the passing of multiple roles to deploy hooks;
they can only look at the instance's first role. Doesn't affect
appcloud, but needs to be fixed for xcloud.