Don't always operate on localhost.
This is in preparation for integrating an app on a new app server;
when a new app server is added to a running cluster, the /data/$app
directory is the same, but none of the deploy hooks have run, nor have
bundled gems been installed, nor have the app servers started. We need
to do that on the new app server from the app master without touching
the app master (localhost).