Added integrate task to engineyard-serverside.
This task will be invoked when a new application server is added to a cluster.
It's similar to a deploy (it runs hooks and everything), but does not fetch
any new code so that your application servers stay in sync.

Refactored code pushing; EY::Server no longer knows how to #push_code, but it
does know how to #sync_directory. This is a better level of abstraction.