Compile assets before the migration page is enabled
With Rails 3.1, asset compilation can take a very long time. If you deploy with migrations (-m), the maintenance page would be enabled BEFORE the assets was compiled and the site would be off-line. There is no need for the maintenance page to be enabled until after the assets have been compiled