Downgrade "maintenance page still up" message to info
When this message is a notice, it gets picked up by the EY Cloud
dashboard deploy log parser. Since EY Cloud doesn't know about
maintenance page status, it's very difficult to clear the message from
the dashboard when the maintenance page is taken down. This leads to
confusion when the dashboard says "maintenance page still up".

This message always displays at the very end of the deploy, so it
should be noticable, even without the notice styling.