Add option to silence missing Gemfile.lock warning
This config option will allow warning-less deploys on Cloud when not
using a version controlled Gemfile.lock

When developing an application on Windows or with multiple ruby engines
you cannot rely on a single Gemfile.lock as it will resolve differently
based on those factors. In these cases you'll want to pessimistically
lock versions in the Gemfile and not version control your Gemfile.lock,
resulting in extra time during deploy to re-resolve the dependencies.