Make sure ey.yml is loaded during the rollback action.
Fixes github issue #87

This was intended all along, as is evidenced by the test that
verifies that rollback works with a complex ey.yml file. The test did
not actually verify that ey.yml was loaded at all. A new test has been
added that verifies that the config is loaded.