Maintenance still up notice => status for visibility without stickiness
It's confusing when the maintenance page still up notice is posted on
the dashboard and sticks around after the maintenance page has been
taken down. The work has already been done in the dashboard to treat
maintenance requests as first class citizens, so this could be solved in
a better way soon.