Supports *experimental* ey.yml option `sync_assets`
When set to true, asset compilation will be executed on the primary
deployment server and rsync'd to other servers.

This feature is experimental.

It has been tested live but the current approach may not prove to be the
most effective way to speed up asset compilation. Future versions will
attempt to maintain a similar feature, but the implementation is subject
to change.