Run configure command provided by the server that confirms app config
If the command does not exist on the server, don't run it and continue
as usual. If the command exists, then run it and fail if it fails.

This allows the server to specify any sort of configuration loading and
application checking that it needs to do so extra features can be added
without extending serverside and synchronizing versions on the server
with features provided by the server.