Adjust how maintenance page is generated
Previously we used `cp` to copy the maintenance page to the correct
location. This can fail in cases where engineyard-serverside is not
installed (or a different version is installed) on the non-deploying
server. (serverside runs on the app_master and sends it's commands out
to the other servers)

Now, instead, we use `echo >` with the full contents of maintenance page.