Threw down a failure handler stub
For the moment, just so the failure handler is in place, we're
basically considering all deploy failures to be fatal problems
with no reversion.

This will allow us to test the happy path in a real environment,
and it will also give us a bit of time to think about the way we
want to actually handle failures.