Extracted callback logic out of DeployBase
In an effort to make the service hooks implementation easier
(or at least possible), this extracts the callback-related
logic out of `EY::Serverside::DeployBase` and drops it into
its very own subsystem, `EY::Serverside::Callbacks`.

This new subsystem is responsible for both the loading and the
dispatching of detected hooks for the various callbacks. The
primary surface area is this:

* `EY::Serverside::Callbacks.load` takes an instance of
  `EY::Serverside::Paths`, uses it to detect deploy hooks,
  and returns an `EY::Serverside::Callbacks::Collection` that
  wraps the detected hooks.
* `EY::Serverside::Callbacks::Collection#execute` takes a "runner"
  (primarily an instance of `EY::Serverside::DeployBase`) and a
  callback name (`:before_deploy`, etc), then executes all hooks
  matching that callback.