Ensures a single ruby distribute call
As Ruby hooks are executed en masse remotely during a single
`engineyard-serverside hook hook_name` call, and we can't
reasonably provide a path to a ruby hook file for that call,
we need to basically either distribute one ruby hook or zero
ruby hooks (depending on whether or not there are *any* to
distribute).

To that end, the base #distribute method for Collection now
filters the list of matching hooks before it passes the
hooks to Distributor such that all but the first ruby hook
in the list is removed.