Initial executable service hook wrapper
This is largely the same as the executable deploy hook
wrapper. The primary difference is that instead of just the hook
name, the argument for this script should be of the form
`service_name/hook_name`. Aside from that, the only real
difference is the base location for loading the hook in question.