First pass at totally remote hook running
Johann discovered that the way we were diestributing hooks is
a bit off. To that end, we're changing things up just a bit, in
so far as instead of running executable hooks directly and
ruby hooks by `hook` invocation, we're going to do all of them
via `hook` invocation.

Since we don't have a way to run the tests locally at present,
this commit is a bit premature, but we need the CI output.