Stop filtering before hook distribution
For service hooks to work as intended, we need to *not* do
existence checks on the app master end of the equation.

Instead, we need to just distribute the callback call to all
applicable targets and let the target decide if it's going to
run a hook.

An interesting ans possibly really handy side-effect of this is
that every callback will have a related log on every applicable
instance in a given environment.