Cleaned up the rakefile
Using a newer RSpec than previously means a few things, some good,
some not so good, and some awesome.

Filed into the "awesome" category is that we don't have to do the
whole `spec/` symlink song and dance anymore. Instead, we just
configure a task for both the "inside" and "outside" suites,
specifying the directory in question, and let 'er rip.'

For giggles, even though we're not writing integration tests just
yet, I've also added in the `cucucmber` task.