Tweaked Server#role example
It looks like `Set` is implemented differently between ruby 1.8.7
and ruby 1.9+, to the point that we can't guarantee that `Set#to_a`
will return the items in the same order that we provided them.

To that end, rather than checking to ensure that it returns the
first role that we provided, we have to make sure that it returns
the first role in the set.