More shell spec tweaks for 1.8.7
Generally, I prefer to use `Object.new` for my test doubles. This
presents a problem when stubbing methods that are marked as
private for `Object` on 1.8.7, like `gsub`.

That being the case, more or less everything that receives `gsub`
in the `EY::Serverside::Shell` spec has been switched to a string.